fixing

+++ install_os/install_things.sh
@@ -9,17 +9,19 @@ check_presence()
 	local ISPRESENT=$1
 	local SHOULDBE=$2
 	local FILE=$3
+
 	grep -q "$ISPRESENT" $FILE || (echo $SHOULDBE >> $FILE \
 		&& echo "Added \"$SHOULDBE\" to \"$FILE\"")
 }
 
-mkdir -p /home/kgotfryd/bin
+check_presence "export VISUAL" "export VISUAL=vim" ~/.bashrc
+check_presence "export EDITOR" "export EDITOR=\"\$VISUAL\"" ~/.bashrc
 
-check_presence "ordinary_backup/scripts" "export PATH=/home/kgotfryd/bin:/home/kgotfryd/programming/workspace/ordinary_backup/scripts:$PATH" ~/.bashrc
+check_presence "ordinary_backup/scripts" "export PATH=/home/kgotfryd/programming/workspace/ordinary_backup/scripts:$PATH" ~/.bashrc
 check_presence "cdw()" "cdw() { cd /home/kgotfryd/programming/workspace; }" ~/.bashrc
 check_presence "cdwl()" "cdwl() { cdw; cd linux; }" ~/.bashrc
 check_presence "cdwo()" "cdwo() { cdw; cd ordinary_backup; }" ~/.bashrc
-check_presence "c()" "c() { if [[ \"\$1\" =~ ^[.]+$ ]]; then NUM=\$(echo \"\$1\" | awk -F\".\" \'{print NF-1}\'); for i in \$( seq 1 \$NUM ); do cd .. ; done; else cd \"\$*\" ; fi ; l ; }" ~/.bashrc
+check_presence "c()" "c() { if [[ \"\$1\" =~ ^[.]+$ ]]; then NUM=\$(echo \"\$1\" | awk -F\".\" '{print NF-1}'); for i in \$( seq 1 \$NUM ); do cd .. ; done; else cd \"\$*\" ; fi ; l ; }" ~/.bashrc
 
 check_presence "e()" "e() { exit; }" ~/.bashrc
 check_presence "f()" "f() { find ./ -name "*$1"; }" ~/.bashrc
@@ -38,8 +40,11 @@ fi
 
 
 #vim installation
-ln -s $REPOPATH/vim/vimrc ~/.vimrc
-read -p "Install vim from submodule? y/n: "
+if [ ! -f ~/.vimrc ]; then
+	ln -s $REPOPATH/vim/vimrc ~/.vimrc
+fi
+
+read -p "Install vim from submodule? y/n: " REPLY
 if [ $REPLY == "y" ]; then
 	echo "installing vim"
 	#fetch all submodules
@@ -54,7 +59,7 @@ if [ $REPLY == "y" ]; then
 	popd
 fi
 
-read -p "Install vundle from github? y/n (rather y):"
+read -p "Install vundle from github? y/n (rather y):" REPLY
 if [ $REPLY = "y" ]; then
 	if [ ! -d ~/.vim/bundle/Vundle.vim ]; then
 		git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
@@ -62,22 +67,33 @@ if [ $REPLY = "y" ]; then
 fi
 
 #clone linux repo
-read -p "Install linux from submodule? y/n: "
+read -p "Install linux from submodule? y/n: " REPLY
 if [ $REPLY == "y" ]; then
 	if [ ! -d ~/programming/workspace/linux ]; then
 		git clone https://github.com/torvalds/linux.git ~/programming/workspace/linux
 	fi
 fi
 
-read -p "Install universal ctags? y/n: "
+read -p "Install universal ctags? y/n: " REPLY
 if [ $REPLY == "y" ]; then
-	tempdir = `mktemp -d`
+
+	sudo apt install \
+        gcc make \
+        pkg-config autoconf automake \
+        python3-docutils \
+        libseccomp-dev \
+        libjansson-dev \
+        libyaml-dev \
+        libxml2-dev
+
+	tempdir=`mktemp -d`
 	pushd $tempdir
 	git clone https://github.com/universal-ctags/ctags.git
 	pushd ctags
-	./configure --prefix=/home/kgotfryd
+	./autogen.sh
+	./configure
 	make
-	make install
+	sudo make install
 	popd
 	popd
 	rm -rf $tempdir
